+++ src/Test/Scripts/ServerExecutionProfile.py
@@ -32,7 +32,7 @@ def main():
     parser.add_argument(
         "--model", type=str, nargs="+", help="Model Names", required=True
     )
-    parser.add_argument("--cpus", type=int, help="Number of CPUs", required=True)
+    parser.add_argument("--cpus", type=float, help="Number of CPUs", required=True)
     parser.add_argument("--server", type=str, help="Server Name", required=True)
     parser.add_argument("--runs", type=int, help="Number of Runs", default=1)
